SCB-1612 Rename oas-validator-compliance to oas-validator-style Rename some classes
