+++ oas-validator/oas-validator-style/src/test/java/org/apache/servicecomb/toolkit/oasv/style/validator/OasStyleCheckTestBase.java
@@ -20,7 +20,7 @@ package org.apache.servicecomb.toolkit.oasv.style.validator;
 import io.swagger.v3.parser.core.models.SwaggerParseResult;
 import io.swagger.v3.parser.util.ClasspathHelper;
 import org.apache.servicecomb.toolkit.oasv.FactoryOptions;
-import org.apache.servicecomb.toolkit.oasv.style.ComplianceCheckParser;
+import org.apache.servicecomb.toolkit.oasv.style.StyleCheckParser;
 import org.apache.servicecomb.toolkit.oasv.validation.test.OasValidatorTestBase;
 
 import java.util.Collections;
@@ -30,7 +30,7 @@ public class OasStyleCheckTestBase extends OasValidatorTestBase {
   @Override
   protected SwaggerParseResult parseAbsolute(String classpath) {
     String content = ClasspathHelper.loadFileFromClasspath(classpath);
-    return ComplianceCheckParser.parseYaml(content);
+    return StyleCheckParser.parseYaml(content);
   }
 
   protected FactoryOptions singleOption(String key, String value) {
