+++ oas-validator/oas-validator-web/src/main/resources/templates/validator/style.html
@@ -90,7 +90,7 @@
     $(function(){
     
     $('#complianceValidate').on('click', function(e) {
-      var url = "/api/compliance";
+      var url = "/api/style";
       
       var yaml = $('textarea[name="yaml"]').val(); if (yaml=="") {return;}
